Refactored packages for future growth and changes.
