+++ pkg/exec/README.md
@@ -1,0 +1,28 @@
+
+# exec
+    import "github.com/coralproject/xenia/pkg/exec"
+
+Package exec provides support for executing Sets and their different types
+of commands.
+
+
+
+
+
+
+## func Set
+``` go
+func Set(context interface{}, db *db.DB, set *query.Set, vars map[string]string) *query.Result
+```
+Set executes the specified query set by name.
+
+
+
+
+
+
+
+
+
+- - -
+Generated by [godoc2md](http://godoc.org/github.com/davecheney/godoc2md)
