--- pkg/query/setup_test.go
@@ -1,20 +1,0 @@
-package query_test
-
-import (
-	"github.com/ardanlabs/kit/cfg"
-	"github.com/ardanlabs/kit/db/mongo"
-	"github.com/ardanlabs/kit/tests"
-)
-
-func init() {
-	tests.Init("XENIA")
-
-	cfg := mongo.Config{
-		Host:     cfg.MustString("MONGO_HOST"),
-		AuthDB:   cfg.MustString("MONGO_AUTHDB"),
-		DB:       cfg.MustString("MONGO_DB"),
-		User:     cfg.MustString("MONGO_USER"),
-		Password: cfg.MustString("MONGO_PASS"),
-	}
-	tests.InitMongo(cfg)
-}
